Extract commit-focus spec to sibling file; use toHaveFocus matcher
CodeRabbit findings on PR #585: the harness pushed the main spec past the
800-line Biome limit, and the manual vi.waitFor reimplemented the built-in
retry-able toHaveFocus matcher. Both trivial, both applied.
