+++ app/(app)/app/practice/components/practice-view.browser.spec.tsx
@@ -797,88 +797,3 @@ test('calls onEndSession from the bottom-bar End session button on the last tuto
   expect(onEndSession).toHaveBeenCalledTimes(1);
   expect(onNextQuestion).not.toHaveBeenCalled();
 });
-
-function TutorCommitFocusHarness() {
-  const [submitResult, setSubmitResult] = useState<{
-    attemptId: string;
-    isCorrect: boolean;
-    correctChoiceId: string;
-    explanationMd: string;
-    referenceMd: null;
-    choiceExplanations: [];
-  } | null>(null);
-
-  return (
-    <>
-      <button
-        type="button"
-        onClick={() =>
-          setSubmitResult({
-            attemptId: fixtureQuestion1Id,
-            isCorrect: true,
-            correctChoiceId: fixtureChoiceAId,
-            explanationMd: 'Because.',
-            referenceMd: null,
-            choiceExplanations: [],
-          })
-        }
-      >
-        complete-commit
-      </button>
-      <PracticeView
-        sessionInfo={{
-          sessionId: fixtureSession1Id,
-          mode: 'tutor',
-
-          deadlineAt: null,
-
-          index: 0,
-          total: 3,
-          isMarkedForReview: false,
-        }}
-        loadState={{ status: 'ready' }}
-        question={{
-          questionId: fixtureQuestion1Id,
-          slug: 'question-1',
-          stemMd: 'What is the next best step?',
-          difficulty: 'easy',
-          choices: [
-            {
-              id: fixtureChoiceAId,
-              label: 'A',
-              textMd: 'Option A',
-              sortOrder: 1,
-            },
-          ],
-          session: null,
-        }}
-        selectedChoiceId={fixtureChoiceAId}
-        isAnswered={submitResult !== null}
-        submitResult={submitResult}
-        isPending={false}
-        bookmarkStatus="idle"
-        isBookmarked={false}
-        onTryAgain={() => undefined}
-        onToggleBookmark={() => undefined}
-        onSelectChoice={() => undefined}
-        onNextQuestion={() => undefined}
-      />
-    </>
-  );
-}
-
-test('moves focus to the answer feedback when a commit completes', async () => {
-  const screen = await render(<TutorCommitFocusHarness />);
-
-  await screen.getByRole('button', { name: 'complete-commit' }).click();
-
-  const feedbackRegion = screen.getByTestId('answer-feedback-region');
-  await expect.element(feedbackRegion).toBeInTheDocument();
-  await vi.waitFor(() => {
-    if (document.activeElement !== feedbackRegion.element()) {
-      throw new Error(
-        `Expected focus on the feedback region, got ${document.activeElement?.tagName}`,
-      );
-    }
-  });
-});
